Lib directory for the genann stuff
We might want to build a proper library at soem point, but this is good
enough for now.
